README update + version bump.

+++ src/ExpressiveAnnotations/Toolchain.cs
@@ -62,10 +62,10 @@ namespace ExpressiveAnnotations
             AddFunction<string, bool>("IsUrl", str => str != null && Regex.IsMatch(str, @"^(?:(?:https?|ftp):\/\/)(?:\S+(?::\S*)?@)?(?:(?!(?:10|127)(?:\.\d{1,3}){3})(?!(?:169\.254|192\.168)(?:\.\d{1,3}){2})(?!172\.(?:1[6-9]|2\d|3[0-1])(?:\.\d{1,3}){2})(?:[1-9]\d?|1\d\d|2[01]\d|22[0-3])(?:\.(?:1?\d{1,2}|2[0-4]\d|25[0-5])){2}(?:\.(?:[1-9]\d?|1\d\d|2[0-4]\d|25[0-4]))|(?:(?:[a-z\u00a1-\uffff0-9]-*)*[a-z\u00a1-\uffff0-9]+)(?:\.(?:[a-z\u00a1-\uffff0-9]-*)*[a-z\u00a1-\uffff0-9]+)*(?:\.(?:[a-z\u00a1-\uffff]{2,})))(?::\d{2,5})?(?:\/\S*)?$", RegexOptions.IgnoreCase)); // contributed by Diego Perini: https://gist.github.com/dperini/729294 (https://mathiasbynens.be/demo/url-regex)
             AddFunction<string, string, bool>("IsRegexMatch", (str, regex) => str != null && regex != null && Regex.IsMatch(str, regex));
             AddFunction<string, Guid>("Guid", str => new Guid(str));
-            AddFunction("Min", (Expression<ParamsDelegate<int, int>>)(items => items.Min()));
-            AddFunction("Max", (Expression<ParamsDelegate<int, int>>)(items => items.Max()));            
-            AddFunction("Sum", (Expression<ParamsDelegate<int, int>>)(items => items.Sum()));
-            AddFunction("Average", (Expression<ParamsDelegate<int, double>>)(items => items.Average()));
+            AddFunction("Min", (Expression<ParamsDelegate<double, double>>)(items => items.Min()));
+            AddFunction("Max", (Expression<ParamsDelegate<double, double>>)(items => items.Max()));            
+            AddFunction("Sum", (Expression<ParamsDelegate<double, double>>)(items => items.Sum()));
+            AddFunction("Average", (Expression<ParamsDelegate<double, double>>)(items => items.Average()));
         }        
 
         /// <summary>
